requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
prompt_toolkit>=3.0.0
